// src/app/ficha-de-matricula/models/matricula.ts
export class EncargadoDeRetiro {
    matriculaId: number = 0;
    rut: string = '';
    nombreCompleto: string = '';
    direccion: string = '';
    telefono: string = '';
    celular: string = '';
    email: string = '';
}